Refactor WMT workload
