+++ algorithmic_efficiency/workloads/wmt/wmt_jax/workload.py
@@ -12,25 +12,16 @@ import numpy as np
 import tensorflow as tf
 
 from algorithmic_efficiency import spec
-from algorithmic_efficiency.workloads.wmt.wmt_jax import bleu
-from algorithmic_efficiency.workloads.wmt.wmt_jax import decode
-from algorithmic_efficiency.workloads.wmt.wmt_jax import input_pipeline
+from algorithmic_efficiency.workloads.wmt.workload import WMT
+from algorithmic_efficiency.workloads.wmt import bleu
+from algorithmic_efficiency.workloads.wmt import decode
 from algorithmic_efficiency.workloads.wmt.wmt_jax import models
 
-VOCAB_PATH = "./wmt_256/sentencepiece_model"
-WORKDIR = "./wmt_256"
-
-
-class WMTWorkload(spec.Workload):
-  """A WMT workload."""
+class WMTWorkload(WMT):
+  """WMT Jax workload."""
 
   def __init__(self):
-    self._eval_ds = None
-    self._train_ds = None
-    self._predict_ds = None
-    self._encoder = None
-    self._vocab_size = 32000
-    self._per_device_batch_size = None
+    super().__init__()
     self._train_config = None
     self._eval_config = None
     self._predict_config = None
@@ -77,41 +68,6 @@ class WMTWorkload(spec.Workload):
 
     return loss, normalizing_factor
 
-  def compute_weighted_accuracy(self, logits, targets, weights=None):
-    """Compute weighted accuracy for log probs and targets.
-
-    Args:
-     logits: [batch, length, num_classes] float array.
-     targets: categorical targets [batch, length] int array.
-     weights: None or array of shape [batch, length]
-
-    Returns:
-      Tuple of scalar loss and batch normalizing factor.
-    """
-    if logits.ndim != targets.ndim + 1:
-      raise ValueError("Incorrect shapes. Got shape %s logits and %s targets" %
-                       (str(logits.shape), str(targets.shape)))
-    loss = jnp.equal(jnp.argmax(logits, axis=-1), targets)
-    normalizing_factor = np.prod(logits.shape[:-1])
-    if weights is not None:
-      loss = loss * weights
-      normalizing_factor = weights.sum()
-
-    return loss.sum(), normalizing_factor
-
-  def compute_metrics(self, logits, labels, weights):
-    """Compute summary metrics."""
-    loss, weight_sum = self.compute_weighted_cross_entropy(
-        logits, labels, weights, 0.0)
-    acc, _ = self.compute_weighted_accuracy(logits, labels, weights)
-    metrics = {
-        "loss": loss.sum(),
-        "accuracy": acc,
-        "denominator": weight_sum,
-    }
-    metrics = jax.lax.psum(metrics, axis_name="batch")
-    return metrics
-
   # Primary eval / decode step functions.
   # -----------------------------------------------------------------------------
 
@@ -121,8 +77,8 @@ class WMTWorkload(spec.Workload):
     weights = jnp.where(targets > 0, 1.0, 0.0)
     logits = models.Transformer(config).apply({"params": params}, inputs,
                                               targets)
-
-    return self.compute_metrics(logits, targets, weights)
+    metrics = self.compute_metrics(logits, targets, weights)
+    return jax.lax.psum(metrics, axis_name="batch")
 
   def initialize_cache(self, inputs, max_decode_len, config):
     """Initialize a cache for a given input shape and max decode length."""
@@ -279,69 +235,10 @@ class WMTWorkload(spec.Workload):
       exemplars += f"{sources[n]}\n\n{references[n]}\n\n{predictions[n]}\n\n"
     return exemplars, bleu_score
 
-  def has_reached_goal(self, eval_result: float) -> bool:
-    return eval_result["bleu"] > self.target_value
-
-  def build_input_queue(self, data_rng: jax.random.PRNGKey, split: str,
-                        data_dir: str, batch_size: int):
-    tf.io.gfile.makedirs(WORKDIR)
-    self._per_device_batch_size = batch_size
-    self._train_ds, self._eval_ds, self._predict_ds, self._encoder = input_pipeline.get_wmt_datasets(
-        vocab_size=self._vocab_size,
-        batch_size=jax.local_device_count() * batch_size,
-        reverse_translation=True,
-        vocab_path=VOCAB_PATH)
-    self._vocab_size = int(self._encoder.vocab_size())
-    return iter(self._train_ds)
-
   @property
   def param_shapes(self):
     init_params, _ = self.init_model_fn(jax.random.PRNGKey(0))
     return jax.tree_map(lambda x: spec.ShapeTuple(x.shape), init_params)
-
-  @property
-  def target_value(self):
-    return 25
-
-  @property
-  def loss_type(self):
-    return spec.LossType.SOFTMAX_CROSS_ENTROPY
-
-  @property
-  def num_train_examples(self):
-    return 5906184
-
-  @property
-  def num_eval_examples(self):
-    return 3004
-
-  @property
-  def train_mean(self):
-    return 0.0
-
-  @property
-  def train_stddev(self):
-    return 1.0
-
-  def model_params_types(self):
-    pass
-
-  @property
-  def max_allowed_runtime_sec(self):
-    return 80000
-
-  @property
-  def eval_period_time_sec(self):
-    return 800
-
-  def _decode_tokens(self, toks):
-    valid_toks = toks[:np.argmax(toks == decode.EOS_ID) + 1].astype(np.int32)
-    return self._encoder.detokenize(valid_toks).numpy().decode("utf-8")
-
-  # Return whether or not a key in spec.ParameterContainer is the output layer
-  # parameters.
-  def is_output_params(self, param_key: spec.ParameterKey) -> bool:
-    pass
 
   def preprocess_for_train(self, selected_raw_input_batch: spec.Tensor,
                            selected_label_batch: spec.Tensor,
@@ -426,11 +323,6 @@ class WMTWorkload(spec.Workload):
                                                   weights)
 
     return loss
-
-  def output_activation_fn(self, logits_batch: spec.Tensor,
-                           loss_type: spec.LossType) -> spec.Tensor:
-    """Return the final activations of the model."""
-    pass
 
   def eval_model(self, params: spec.ParameterContainer,
                  model_state: spec.ModelAuxiliaryState, rng: spec.RandomState,
